modify binary_search.cpp

+++ C++/lib/binary_search.cpp
@@ -1,7 +1,7 @@
 #include <vector>
 
 // binary_search 二分探索
-int binary_searching(std::vector<long long> &str, long long key, long long imin, long long imax) {
+long long binary_searching(std::vector<long long> &str, long long key, long long imin, long long imax) {
     if (imax < imin) return -1;
     long long imid = imin + (imax - imin) / 2;
     if (str.at(imid) < key) {
